fix type to printfs. (camel_imap4_command_newv, unexpected_token):
2005-08-22  Not Zed  <NotZed@Ximian.com>

        * camel-imap4-utils.c (camel_imap4_utils_set_unexpected_token_error):

        * camel-imap4-stream.c (camel_imap4_stream_next_token):

        * camel-imap4-command.c (imap4_command_append_string): fix type to
        printfs.
        (camel_imap4_command_newv, unexpected_token):

+++ camel/providers/imap4/camel-imap4-command.c
@@ -157,14 +157,14 @@ imap4_command_append_string (CamelIMAP4Engine *engine, CamelIMAP4CommandPart **t
 	case IMAP4_STRING_LITERAL:
 		if (engine->capa & CAMEL_IMAP4_CAPABILITY_LITERALPLUS) {
 			/* we have to send a literal, but the server supports LITERAL+ so use that */
-			g_string_append_printf (str, "{%u+}\r\n%s", strlen (string), string);
+			g_string_append_printf (str, "{%u+}\r\n%s", (unsigned int)strlen(string), string);
 		} else {
 			/* we have to make it a literal */
 			literal = g_new (CamelIMAP4Literal, 1);
 			literal->type = CAMEL_IMAP4_LITERAL_STRING;
 			literal->literal.string = g_strdup (string);
 			
-			g_string_append_printf (str, "{%u}\r\n", strlen (string));
+			g_string_append_printf (str, "{%u}\r\n", (unsigned int)strlen(string));
 			
 			(*tail)->buffer = g_strdup (str->str);
 			(*tail)->buflen = str->len;
@@ -253,7 +253,7 @@ camel_imap4_command_newv (CamelIMAP4Engine *engine, CamelIMAP4Folder *imap4_fold
 				
 				/* FIXME: take advantage of LITERAL+? */
 				len = camel_imap4_literal_length (literal);
-				g_string_append_printf (str, "{%u}\r\n", len);
+				g_string_append_printf (str, "{%u}\r\n", (unsigned int)len);
 				
 				tail->buffer = g_strdup (str->str);
 				tail->buflen = str->len;
@@ -491,7 +491,7 @@ unexpected_token (camel_imap4_token_t *token)
 	        fprintf (stderr, "\"%s\"", token->v.qstring);
 		break;
 	case CAMEL_IMAP4_TOKEN_LITERAL:
-		fprintf (stderr, "{%u}", token->v.literal);
+		fprintf (stderr, "{%u}", (unsigned int)token->v.literal);
 		break;
 	default:
 		fprintf (stderr, "%c", (unsigned char) (token->token & 0xff));
